Major version suffixes are not allowed at major versions v0 or v1
https://golang.org/ref/mod#major-version-suffixes

Major version suffixes are not allowed at major versions v0 or v1. There is
no need to change the module path between v0 and v1 because v0 versions are
unstable and have no compatibility guarantee. Additionally, for most modules,
v1 is backwards compatible with the last v0 version; a v1 version acts as a
commitment to compatibility, rather than an indication of incompatible
changes compared with v0.
