+++ go.mod
@@ -3,10 +3,8 @@ module github.com/Workiva/go-rest
 go 1.15
 
 require (
-	github.com/gorilla/context v1.1.1
-	github.com/gorilla/mux v1.6.2
+	github.com/gorilla/mux v1.8.0
 	github.com/hoisie/mustache v0.0.0-20160804235033-6375acf62c69
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/net v0.0.0-20190813141303-74dc4d7220e7
 	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0
 )
