+++ rest/context_test.go
@@ -18,14 +18,15 @@ package rest
 
 import (
 	"bytes"
+	"context"
 	"crypto/tls"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
-	gContext "github.com/gorilla/context"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 // Test Handlers
@@ -55,6 +56,8 @@ type ComplexTestResourceHandler struct {
 	BaseResourceHandler
 }
 
+type contextKey string
+
 func (t ComplexTestResourceHandler) ResourceName() string {
 	return "resources"
 }
@@ -69,21 +72,76 @@ func (t ComplexTestResourceHandler) CreateResource(r RequestContext, data Payloa
 	return resource, nil
 }
 
+// Ensures that we're correctly setting a value on the http.Request Context
+func TestSetValueOnRequestContext(t *testing.T) {
+	req, err := http.NewRequestWithContext(context.Background(), "GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
+	k := contextKey("mykey")
+	val := req.Context().Value(k)
+	assert.Nil(t, val)
+
+	expected := "myval"
+	req = setValueOnRequestContext(req, k, expected)
+	actual := req.Context().Value(k)
+	assert.Equal(t, expected, actual)
+}
+
+func TestValueReturnsRequest(t *testing.T) {
+	assert := assert.New(t)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
+	ctx := NewContext(req, nil)
+	val := ctx.Value(requestKey)
+	assert.Equal(req, val, "Value called with the requestKey should return the request")
+}
+
+func TestValueReturnsPreviouslySetValue(t *testing.T) {
+	assert := assert.New(t)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
+	ctx := NewContext(req, nil)
+	k := contextKey("mykey")
+	expected := "myval"
+	ctx = ctx.WithValue(k, expected)
+
+	actual := ctx.Value(k)
+	assert.Equal(expected, actual, "Value called with a key should find the value on the request's context if it exists")
+}
+
+func TestValueReturnsNilIfNoKey(t *testing.T) {
+	assert := assert.New(t)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
+	ctx := NewContext(req, nil).(*requestContext)
+	k := contextKey("mykey")
+
+	actual := ctx.Value(k)
+	assert.Nil(actual, "Value called with a non-existent key should return nil")
+}
+
 // Ensures that if a limit doesn't exist on the context, the default is returned.
 func TestLimitDefault(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 	assert.Equal(100, ctx.Limit())
 }
 
 // Ensures that if an invalid limit value is on the context, the default is returned.
 func TestLimitBadValue(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 	ctx = ctx.WithValue(limitKey, "blah")
 	assert.Equal(100, ctx.Limit())
 }
@@ -91,9 +149,11 @@ func TestLimitBadValue(t *testing.T) {
 // Ensures that the correct limit is returned from the context.
 func TestLimit(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 	ctx = ctx.WithValue(limitKey, "5")
 	assert.Equal(5, ctx.Limit())
 }
@@ -101,9 +161,11 @@ func TestLimit(t *testing.T) {
 // Ensures that Messages returns the messages set on the context.
 func TestMessagesNoError(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 	message := "foo"
 
 	assert.Equal(0, len(ctx.Messages()))
@@ -119,12 +181,14 @@ func TestMessagesNoError(t *testing.T) {
 // when an error is set.
 func TestMessagesWithError(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 	message := "foo"
 	errMessage := "blah"
-	err := fmt.Errorf(errMessage)
+	err = fmt.Errorf(errMessage)
 
 	ctx = ctx.setError(err)
 	if assert.Equal(1, len(ctx.Messages())) {
@@ -142,9 +206,11 @@ func TestMessagesWithError(t *testing.T) {
 // Ensures that Header returns the request Header.
 func TestHeader(t *testing.T) {
 	assert := assert.New(t)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", nil)
+	req, err := http.NewRequest("GET", "http://example.com/foo", nil)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 
 	assert.Equal(req.Header, ctx.Header())
 }
@@ -154,9 +220,11 @@ func TestBody(t *testing.T) {
 	assert := assert.New(t)
 	payload := []byte(`[{"foo": "bar"}]`)
 	r := bytes.NewReader(payload)
-	req, _ := http.NewRequest("GET", "http://example.com/foo", r)
+	req, err := http.NewRequest("GET", "http://example.com/foo", r)
+	require.NoError(t, err)
+
 	writer := httptest.NewRecorder()
-	ctx := NewContext(nil, req, writer)
+	ctx := NewContext(req, writer)
 
 	assert.Equal(payload, ctx.Body().Bytes())
 }
@@ -168,30 +236,36 @@ func TestBuildURL(t *testing.T) {
 	api.RegisterResourceHandler(TestResourceHandler{})
 	api.RegisterResourceHandler(ComplexTestResourceHandler{})
 
-	req, _ := http.NewRequest("GET", "https://example.com/api/v1/widgets", nil)
-	gContext.Set(req, "version", "1")
+	req, err := http.NewRequest("GET", "https://example.com/api/v1/widgets", nil)
+	require.NoError(t, err)
+	req = setValueOnRequestContext(req, "version", "1")
 
 	writer := httptest.NewRecorder()
-	ctx := NewContextWithRouter(nil, req, writer, api.(*muxAPI).router)
+	ctx := NewContextWithRouter(req, writer, api.(*muxAPI).router)
 
-	url, _ := ctx.BuildURL("widgets", HandleCreate, nil)
+	url, err := ctx.BuildURL("widgets", HandleCreate, nil)
+	require.NoError(t, err)
 	assert.Equal(url.String(), "http://example.com/api/v1/widgets")
 
-	url, _ = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "111"})
+	url, err = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "111"})
+	require.NoError(t, err)
 	assert.Equal(url.String(), "http://example.com/api/v1/widgets/111")
 
-	url, _ = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "111"})
+	url, err = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "111"})
+	require.NoError(t, err)
 	assert.Equal(url.Path, "/api/v1/widgets/111")
 
 	// Secure request should produce https URL
 	req.TLS = &tls.ConnectionState{}
-	url, _ = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "222"})
+	url, err = ctx.BuildURL("widgets", HandleRead, RouteVars{"resource_id": "222"})
+	require.NoError(t, err)
 	assert.Equal(url.String(), "https://example.com/api/v1/widgets/222")
 
 	// Make sure this works with another version number
-	gContext.Set(req, "version", "2")
-	url, _ = ctx.BuildURL("resources", HandleCreate, RouteVars{
+	ctx = ctx.WithValue("version", "2")
+	url, err = ctx.BuildURL("resources", HandleCreate, RouteVars{
 		"company":  "acme",
 		"category": "anvils"})
+	require.NoError(t, err)
 	assert.Equal(url.String(), "https://example.com/api/v2/acme/anvils/resources")
 }
