+++ rest/handler.go
@@ -137,7 +137,7 @@ type requestHandler struct {
 // The serialization mechanism used is specified by the "format" query parameter.
 func (h requestHandler) handleCreate(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
 
@@ -178,7 +178,7 @@ func (h requestHandler) handleCreate(handler ResourceHandler) http.Handler {
 // serialization mechanism used is specified by the "format" query parameter.
 func (h requestHandler) handleReadList(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
 
@@ -206,7 +206,7 @@ func (h requestHandler) handleReadList(handler ResourceHandler) http.Handler {
 // mechanism used is specified by the "format" query parameter.
 func (h requestHandler) handleRead(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
 
@@ -229,7 +229,7 @@ func (h requestHandler) handleRead(handler ResourceHandler) http.Handler {
 // parameter.
 func (h requestHandler) handleUpdateList(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
 
@@ -278,7 +278,7 @@ func (h requestHandler) handleUpdateList(handler ResourceHandler) http.Handler {
 // parameter.
 func (h requestHandler) handleUpdate(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
 
@@ -313,7 +313,7 @@ func (h requestHandler) handleUpdate(handler ResourceHandler) http.Handler {
 // mechanism used is specified by the "format" query parameter.
 func (h requestHandler) handleDelete(handler ResourceHandler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := NewContextWithRouter(nil, r, w, h.router)
+		ctx := NewContextWithRouter(r, w, h.router)
 		version := ctx.Version()
 		rules := handler.Rules()
