--- v1/go.mod
@@ -1,10 +1,0 @@
-module github.com/Workiva/go-rest/v1
-
-go 1.15
-
-require (
-	github.com/gorilla/mux v1.8.0
-	github.com/hoisie/mustache v0.0.0-20160804235033-6375acf62c69
-	github.com/stretchr/testify v1.6.1
-	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0
-)
